Refactoring: GridCreatingConditionTriangle::AddRemeshPolygonCommand implementation is separated from gridcreatingconditiontriangle.cpp.
